File tests passing

+++ licensing/python_tests/licensing_utils.py
@@ -3,7 +3,9 @@ import os
 import pandas as pd
 
 
-def this_should_require_a_full_license_udt():
+def this_should_require_a_full_license_udt(
+    test_load_save=True, n_target_classes=2, num_data_points=2
+):
     from thirdai import bolt
 
     model = bolt.UniversalDeepTransformer(
@@ -12,20 +14,27 @@ def this_should_require_a_full_license_udt():
             "col_2": bolt.types.categorical(),
         },
         target="col_1",
-        n_target_classes=2,
+        n_target_classes=n_target_classes,
     )
 
-    df = pd.DataFrame({"col_1": [0.0, 1.0], "col_2": [0, 1]})
+    df = pd.DataFrame(
+        {
+            "col_1": [i % n_target_classes for i in range(num_data_points)],
+            "col_2": [i % n_target_classes for i in range(num_data_points)],
+        }
+    )
     df.to_csv("temp_training.csv")
 
     model.train("temp_training.csv")
 
-    model.save("temp_save_loc")
+    if test_load_save:
+        model.save("temp_save_loc")
 
-    bolt.UniversalDeepTransformer.load("temp_save_loc")
+        bolt.UniversalDeepTransformer.load("temp_save_loc")
+
+        os.remove("temp_save_loc")
 
     os.remove("temp_training.csv")
-    os.remove("temp_save_loc")
 
 
 LOCAL_HEARTBEAT_SERVER = f"http://localhost:50421"
